Move third-party code into subprojects directory.
