+++ src/device/r4300/mips_instructions.def
@@ -56,8 +56,9 @@
 #include "device/memory/memory.h"
 #include "device/rcp/mi/mi_controller.h"
 #include "device/rdram/rdram.h"
-#include "main/xxHash/xxhash.h"
 #include "osal/preproc.h"
+
+#include <xxhash.h>
 
 #include <inttypes.h>
 #include <stdint.h>
