+++ subprojects/minizip/crypt.h
